requirements.txt
sentence-transformers==2.2.2
langchain
flask
pypdf
python-dotenv
pinecone
langchain-pinecone
langchain_community
langchain_openai
langchain_experimental
ollama
-e .